Major Update of Core Algorithm
This update performs a major overhaul of the code which should improve
the quality of the results and makes the code for the generation of the
burgers and line vectors more readable.

All changes are made in the backend and there is nothing different
involved in running the program.

Specifically, the code now directly relates the components of the
lattice curvature tensor, Kappa, to the dislocation densities for the
various dislocation types.  Previously, 5 components of the Nye tensor
were being calculated from the 6 components of the curvature tensor,
resulting in a losee of information.  By using the curvature tensor
directly, we take advantage of all 6 available components from the EBSD
measurements.  This will result in calculated GND density values that
are significantly higher than the previous versions of the code, and
should be nearer to the physical values.