Removed C++ dependencies
Previously, the lattice curvature calculations were done in c++ for
speed.  This has been eliminated and now all calculations are done in
Matlab/MTEX with no noticeable decrease in performance.   This should
make it much easier to add support for other cyrstal symmetries in the
future.  Also made small updates to the readme file, including a link to
a series of tutorials on the code by Chris Cochrane.